rename(back): instruction -> profile로 도메인 이름 변경
- instruction는 너무 추상적이어서 다른 도메인과 헷갈릴 수 있기 때문에 구체적인 profile로 대체
