--- back/aboutme-adapter/src/main/kotlin/com/aboutme/adapter/in/http/instruction/controller/InstructionController.kt
@@ -1,44 +1,0 @@
-package com.aboutme.adapter.`in`.http.instruction.controller
-
-import com.aboutme.adapter.`in`.http.instruction.controller.api.InstructionApi
-import com.aboutme.adapter.`in`.http.instruction.req.InstructionReq
-import com.aboutme.app.file.service.dto.rep.FileUploadRep
-import com.aboutme.app.instruction.port.`in`.InstructionUseCase
-import com.aboutme.app.instruction.service.dto.rep.InstructionDetailRep
-import org.springframework.http.MediaType
-import org.springframework.web.bind.annotation.DeleteMapping
-import org.springframework.web.bind.annotation.GetMapping
-import org.springframework.web.bind.annotation.PostMapping
-import org.springframework.web.bind.annotation.RequestBody
-import org.springframework.web.bind.annotation.RequestPart
-import org.springframework.web.bind.annotation.RestController
-import org.springframework.web.multipart.MultipartFile
-
-@RestController
-class InstructionController(
-    private val instructionUseCase: InstructionUseCase,
-) : InstructionApi {
-    @PostMapping("/instructions")
-    override fun createOrUpdate(
-        @RequestBody req: InstructionReq,
-    ) {
-        instructionUseCase.createOrUpdate(req.toCommand())
-    }
-
-    @PostMapping(value = ["/instructions/profile-images"], consumes = [MediaType.MULTIPART_FORM_DATA_VALUE])
-    override fun replaceProfileImage(
-        @RequestPart img: MultipartFile,
-    ): FileUploadRep {
-        return instructionUseCase.replaceProfileImage(img)
-    }
-
-    @GetMapping("/instructions")
-    override fun readDetail(): InstructionDetailRep {
-        return instructionUseCase.readDetail()
-    }
-
-    @DeleteMapping("/instructions/profile-images")
-    override fun deleteProfileImage() {
-        instructionUseCase.deleteProfileImage()
-    }
-}
