--- back/aboutme-core/src/main/kotlin/com/aboutme/core/instruction/error/InstructionErrorCode.kt
@@ -1,13 +1,0 @@
-package com.aboutme.core.instruction.error
-
-import com.aboutme.common.exception.payload.BaseErrorCode
-import com.aboutme.common.exception.payload.ReasonCode
-import com.aboutme.externalapi.exception.payload.StatusCode
-
-enum class InstructionErrorCode(
-    override val statusCode: StatusCode,
-    override val reasonCode: ReasonCode,
-    override val message: String,
-) : BaseErrorCode {
-    NOT_FOUND(StatusCode.NOT_FOUND, ReasonCode.REQUESTED_RESOURCE_NOT_FOUND, "이력서가 존재하지 않습니다."),
-}
